Update Nodejs to v16
In order to fix an error when building the static pages in
Gatsby Cloud, we need to be able to override sub-dependency
versions of deepmerge, but this functionality was added in v8
of NPM CLI, which is associated with v16 of Nodejs.

The update to the GQL query is to fix an error when trying to use
a string for an enum value.